requirements.txt
streamlit>=1.28.0
gspread>=5.7.0
google-auth>=2.17.0
google-auth-oauthlib>=0.4.6 